requirements.txt
click==6.7
Flask==0.12.2
Flask-Login==0.4.0
Flask-Uploads==0.2.1
itsdangerous==0.24
Jinja2==2.9.6
ldap3==2.3
MarkupSafe==1.0
pyasn1==0.3.6
PyYAML==3.12
Werkzeug==0.12.2
